Transform all async ast nodes into sync nodes
so that they can be handled identically. Sync and async nodes propagate
taint in exactly the same way.

Awaits are more or less removed:
`return await x()` is converted into `return x()`.

Any AsyncFunctionDef is converted to a FunctionDef.

Same for AsyncFor and AsyncWith.

Using a transformer makes it easier to replace awaits everywhere.
We have lots of places where we do `isinstance(node, ast.Call)` but
for these we want it to function like `isinstance(node, ast.Call) or
(isinstance(node, ast.Await) and isinstance(node.value, ast.Call))`